change tokensRequested datatype to STRING
previously, `tokensRequested` had the datatype INTEGER. this was
problematic because JS cannot handle integers larger than 53 bits,
and the EVM only supports integers. indeed, if a user requested a
a fractional amount of tokens, JS may not have been able to convert
the value with precision. this commit addresses this issue by changing
the datatype to STRING.

this commit also includes some failing tests (skipped for now) to
address issues wrt conversions (see comments for details).

- update proposalSchema, model, and data
- add migration
- add tests