Set request expiration at creation
Previously, we were setting the expiration only when a request was
accepted. It's cleaner logic to set it when the request is created.

Also, instead of setting the expiration to an epoch length after it
is accepted, set it to the start of the second epoch after the one
in which it is created.