First pass at better frontend error-handling
Do not rethrow errors in page components. Instead, catch them and
handle them by displaying toasts (or whatever is appropriate). Any
errors below the page components should be propagated upward to
the page rather than simply being caught. Error handling in the
pages should reset the view as appropriate and print the error to
the console.