handle errors for reverted contract calls
`getWinningSlate`, `incumbent`, and `projectedUnlockedBalance`
revert if the state conditions have not been met. previously we were
not handling api calls which then run these contract calls. this commit
wraps the calls in try/catch and either returns a useful error response
or ignores the exception because the value can be left undefined.